Add Ecosystem wallet pages in teams (#4491)
## Problem solved

Short description of the bug fixed or feature added

<!-- start pr-codex -->

---

## PR-Codex overview
This PR focuses on restructuring ecosystem-related components and pages in the dashboard.

### Detailed summary
- Deleted `redirects.js`.
- Updated page components for ecosystem permissions and creation.
- Added new components for ecosystem layout and header.
- Modified functions for fetching ecosystem data and ecosystem selection.

> The following files were skipped due to too many changes: `apps/dashboard/src/app/(dashboard)/dashboard/connect/ecosystem/[slug]/(active)/components/client/ecosystem-header.client.tsx`, `apps/dashboard/src/app/(dashboard)/dashboard/connect/ecosystem/EcosystemLandingPage.tsx`, `apps/dashboard/src/app/(dashboard)/dashboard/connect/ecosystem/page.tsx`

> ✨ Ask PR-Codex anything about this PR by commenting with `/codex {your question}`

<!-- end pr-codex -->
