+++ apps/dashboard/src/app/(dashboard)/dashboard/connect/ecosystem/[slug]/(active)/page.tsx
@@ -1,0 +1,5 @@
+import { EcosystemPermissionsPage } from "./EcosystemPermissionsPage";
+
+export default function Page({ params }: { params: { slug: string } }) {
+  return <EcosystemPermissionsPage params={params} />;
+}
