+++ apps/dashboard/src/app/team/[team_slug]/[project_slug]/connect/ecosystem/[slug]/(active)/page.tsx
@@ -1,0 +1,5 @@
+import { EcosystemPermissionsPage } from "../../../../../../../(dashboard)/dashboard/connect/ecosystem/[slug]/(active)/EcosystemPermissionsPage";
+
+export default function Page({ params }: { params: { slug: string } }) {
+  return <EcosystemPermissionsPage params={params} />;
+}
