+++ apps/dashboard/src/app/team/[team_slug]/[project_slug]/connect/ecosystem/page.tsx
@@ -1,9 +1,17 @@
-export default function Page() {
+import { EcosystemLandingPage } from "../../../../../(dashboard)/dashboard/connect/ecosystem/EcosystemLandingPage";
+
+export default async function Page(props: {
+  params: { team_slug: string; project_slug: string };
+}) {
+  const { team_slug, project_slug } = props.params;
   return (
-    <div className="h-full py-6 container flex items-center justify-center">
-      <h1 className="text-4xl tracking-tighter text-muted-foreground">
-        Ecosystem Wallets
-      </h1>
+    <div className="max-sm:pt-6">
+      <EcosystemLandingPage
+        ecosystemLayoutPath={`/team/${team_slug}/${project_slug}/connect/ecosystem`}
+      />
     </div>
   );
 }
+
+// because cookies() is used
+export const dynamic = "force-dynamic";
